Separate certificate metric related tests from ServerTest
