+++ core/src/test/java/com/linecorp/armeria/server/ServerTest.java
@@ -23,19 +23,14 @@ import static org.awaitility.Awaitility.await;
 
 import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.PrintWriter;
 import java.lang.management.ManagementFactory;
 import java.lang.management.ThreadInfo;
 import java.net.Socket;
 import java.nio.charset.StandardCharsets;
-import java.security.cert.CertificateException;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Calendar;
-import java.util.Collection;
-import java.util.Date;
 import java.util.List;
 import java.util.Queue;
 import java.util.concurrent.CompletableFuture;
@@ -56,7 +51,6 @@ import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.util.EntityUtils;
-import org.assertj.core.api.AbstractDoubleAssert;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -80,7 +74,6 @@ import com.linecorp.armeria.common.util.Exceptions;
 import com.linecorp.armeria.common.util.ThreadFactories;
 import com.linecorp.armeria.common.util.TimeoutMode;
 import com.linecorp.armeria.internal.common.metric.MicrometerUtil;
-import com.linecorp.armeria.internal.common.util.SelfSignedCertificate;
 import com.linecorp.armeria.internal.testing.AnticipatedException;
 import com.linecorp.armeria.server.logging.AccessLogWriter;
 import com.linecorp.armeria.server.logging.LoggingService;
@@ -101,9 +94,6 @@ class ServerTest {
     private static final long processDelayMillis = 1000;
     private static final long requestTimeoutMillis = 500;
     private static final long idleTimeoutMillis = 500;
-
-    private static final String CERT_VALIDITY_GAUGE_NAME = "armeria.server.certificate.validity";
-    private static final String CERT_VALIDITY_DAYS_GAUGE_NAME = "armeria.server.certificate.validity.days";
 
     private static final EventExecutorGroup asyncExecutorGroup = new DefaultEventExecutorGroup(1);
 
@@ -524,142 +514,6 @@ class ServerTest {
     }
 
     @Test
-    void noTlsMetricGivenNoTlsSetup() {
-        final MeterRegistry meterRegistry = PrometheusMeterRegistries.newRegistry();
-        Server.builder()
-                .service("/", (ctx, req) -> HttpResponse.of(200))
-                .meterRegistry(meterRegistry)
-                .build();
-
-        final Gauge expirationGauge = meterRegistry.find(CERT_VALIDITY_GAUGE_NAME).gauge();
-        final Gauge timeToExpireGauge = meterRegistry.find(CERT_VALIDITY_DAYS_GAUGE_NAME).gauge();
-
-        assertThat(expirationGauge).isNull();
-        assertThat(timeToExpireGauge).isNull();
-    }
-
-    @Test
-    void tlsMetricGivenConfiguredCertificateNotExpired() throws CertificateException {
-        final SelfSignedCertificate ssc = new SelfSignedCertificate("localhost");
-        final MeterRegistry meterRegistry = PrometheusMeterRegistries.newRegistry();
-        Server.builder()
-              .service("/", (ctx, req) -> HttpResponse.of(200))
-              .meterRegistry(meterRegistry)
-              .tls(ssc.certificate(), ssc.privateKey())
-              .build();
-
-        assertThatGauge(meterRegistry, CERT_VALIDITY_GAUGE_NAME, "localhost").isOne();
-        assertThatGauge(meterRegistry, CERT_VALIDITY_DAYS_GAUGE_NAME, "localhost").isPositive();
-    }
-
-    @Test
-    void tlsMetricGivenSelfSignCertificateNotExpired() {
-        final MeterRegistry meterRegistry = PrometheusMeterRegistries.newRegistry();
-        Server.builder()
-              .service("/", (ctx, req) -> HttpResponse.of(200))
-              .meterRegistry(meterRegistry)
-              .tlsSelfSigned()
-              .build();
-
-        final Gauge expirationGauge = meterRegistry.find(CERT_VALIDITY_GAUGE_NAME).gauge();
-        assertThat(expirationGauge).isNotNull();
-        assertThat(expirationGauge.value()).isOne();
-
-        final Gauge timeToExpireGauge = meterRegistry.find(CERT_VALIDITY_DAYS_GAUGE_NAME).gauge();
-        assertThat(timeToExpireGauge).isNotNull();
-        assertThat(timeToExpireGauge.value()).isPositive();
-    }
-
-    @Test
-    void tlsMetricGivenVirtualHostCertificateNotExpired() throws CertificateException {
-        final String defaultHostName = "virtual.com";
-        final String hostnamePattern = "*.virtual.com";
-        final SelfSignedCertificate ssc = new SelfSignedCertificate(hostnamePattern);
-        final MeterRegistry meterRegistry = PrometheusMeterRegistries.newRegistry();
-
-        Server.builder()
-              .service("/", (ctx, req) -> HttpResponse.of(200))
-              .meterRegistry(meterRegistry)
-              .tls(ssc.certificate(), ssc.privateKey())
-              .virtualHost(defaultHostName,hostnamePattern)
-              .service("/", (ctx, req) -> HttpResponse.of(200))
-              .tlsSelfSigned().and()
-              .build();
-
-        final Collection<Gauge> validityGauges = meterRegistry.find(CERT_VALIDITY_GAUGE_NAME).gauges();
-        final Collection<Gauge> daysValidityGauges = meterRegistry.find(CERT_VALIDITY_DAYS_GAUGE_NAME).gauges();
-        // One gauge set for defaultVirtualHost and another for *.virtual.com
-        assertThat(validityGauges.size()).isEqualTo(2);
-        assertThat(daysValidityGauges.size()).isEqualTo(2);
-
-        assertThat(meterRegistry.find(CERT_VALIDITY_GAUGE_NAME)
-                                .tag("common_name", hostnamePattern)
-                                .tag("hostname", defaultHostName)
-                                .gauge().value()).isOne();
-        assertThat(meterRegistry.find(CERT_VALIDITY_DAYS_GAUGE_NAME)
-                                .tag("common_name", hostnamePattern)
-                                .tag("hostname", defaultHostName)
-                                .gauge().value()).isPositive();
-    }
-
-    @Test
-    void tlsMetricGivenCertificateChainNotExpired() {
-        final InputStream expiredCertificateChain = getClass().getResourceAsStream("certificate-chain.pem");
-        final InputStream pk = getClass().getResourceAsStream("pk.key");
-
-        final MeterRegistry meterRegistry = PrometheusMeterRegistries.newRegistry();
-        Server.builder()
-              .service("/", (ctx, req) -> HttpResponse.of(200))
-              .meterRegistry(meterRegistry)
-              .tls(expiredCertificateChain, pk)
-              .build();
-
-        assertThatGauge(meterRegistry, CERT_VALIDITY_GAUGE_NAME, "localhost").isOne();
-        assertThatGauge(meterRegistry, CERT_VALIDITY_DAYS_GAUGE_NAME, "localhost").isPositive();
-        assertThatGauge(meterRegistry, CERT_VALIDITY_GAUGE_NAME, "test.root.armeria").isOne();
-        assertThatGauge(meterRegistry, CERT_VALIDITY_DAYS_GAUGE_NAME, "test.root.armeria").isPositive();
-    }
-
-    @Test
-    void tlsMetricGivenExpired() throws CertificateException {
-        final Calendar calendar = Calendar.getInstance();
-        calendar.add(Calendar.DAY_OF_MONTH, -1);
-        final Date notAfter = calendar.getTime();
-        calendar.add(Calendar.DAY_OF_MONTH, -1);
-        final Date notBefore = calendar.getTime();
-        final SelfSignedCertificate ssc = new SelfSignedCertificate("localhost", notBefore, notAfter);
-
-        final MeterRegistry meterRegistry = PrometheusMeterRegistries.newRegistry();
-        Server.builder()
-              .service("/", (ctx, req) -> HttpResponse.of(200))
-              .meterRegistry(meterRegistry)
-              .tls(ssc.certificate(), ssc.privateKey())
-              .build();
-
-        assertThatGauge(meterRegistry, CERT_VALIDITY_GAUGE_NAME, "localhost").isZero();
-        assertThatGauge(meterRegistry, CERT_VALIDITY_DAYS_GAUGE_NAME, "localhost").isEqualTo(-1);
-    }
-
-    @Test
-    void tlsMetricGivenCertificateChainExpired() {
-        final InputStream expiredCertificateChain = getClass()
-                .getResourceAsStream("expired-certificate-chain.pem");
-        final InputStream pk = getClass().getResourceAsStream("expire-pk.key");
-
-        final MeterRegistry meterRegistry = PrometheusMeterRegistries.newRegistry();
-        Server.builder()
-              .service("/", (ctx, req) -> HttpResponse.of(200))
-              .meterRegistry(meterRegistry)
-              .tls(expiredCertificateChain, pk)
-              .build();
-
-        assertThatGauge(meterRegistry, CERT_VALIDITY_GAUGE_NAME, "localhost").isZero();
-        assertThatGauge(meterRegistry, CERT_VALIDITY_DAYS_GAUGE_NAME, "localhost").isEqualTo(-1);
-        assertThatGauge(meterRegistry, CERT_VALIDITY_GAUGE_NAME, "test.root.armeria").isOne();
-        assertThatGauge(meterRegistry, CERT_VALIDITY_DAYS_GAUGE_NAME, "test.root.armeria").isPositive();
-    }
-
-    @Test
     void blockUntilShutdown() throws Exception {
         final AtomicBoolean stopped = new AtomicBoolean();
         final Server server = Server.builder()
@@ -759,12 +613,5 @@ class ServerTest {
         private void recordThread() {
             threads.add(Thread.currentThread());
         }
-    }
-
-    private static AbstractDoubleAssert<?> assertThatGauge(MeterRegistry meterRegistry, String gaugeName,
-                                                           String cn) {
-        final Gauge gauge = meterRegistry.find(gaugeName).tag("common_name", cn).gauge();
-        assertThat(gauge).isNotNull();
-        return assertThat(gauge.value());
     }
 }
